Move rest of spin functions into spin.go in sensor base (#2981)
